feat: harden ride workflows, authentication, payments, and persistence
Add rotating refresh-token sessions with reuse detection and secure cookies.
Enforce role-based security and consistent exception responses.

Add optimistic locking, Flyway migrations, and PostGIS indexes.

Harden ride lifecycle, driver availability, OTP, validation, and pagination.

Record wallet transactions and platform commission using BigDecimal.

Add admin bootstrap, production configuration, logging, OSRM timeouts, tests, and comprehensive documentation.

+++ src/main/java/com/abhilash/rideops/strategies/impl/WalletPaymentStrategy.java
@@ -33,9 +33,7 @@ public class WalletPaymentStrategy implements PaymentStrategy {
         Driver driver=payment.getRide().getDriver();
         Rider rider=payment.getRide().getRider();
         BigDecimal platformCommission = calculatePlatformCommission(payment.getAmount());
-        double driversCut = BigDecimal.valueOf(payment.getAmount())
-                .subtract(platformCommission)
-                .doubleValue();
+        BigDecimal driversCut = payment.getAmount().subtract(platformCommission);
         walletService.deductMoneyFromWallet(rider.getUser(),payment.getAmount(),null
                 ,payment.getRide(), TransactionMethod.RIDE);
         walletService.addMoneyToWallet(driver.getUser(),driversCut,null
